+++ blocks/form/src/messages.tsx
@@ -1,6 +1,7 @@
 export default {
   invalid: 'This value is invalid',
   emptyFileLabel: ' ',
+  error: 'One of the requirements of this form is invalid.',
   submit: 'Submit',
   unsupported: 'This file type is not supported',
 };