+++ packages/utils/src/validateAppDefinition.test.ts
@@ -114,7 +114,7 @@ describe('checkBlocks', () => {
     expect(error).toBeInstanceOf(AppsembleValidationError);
     expect(error.message).toBe('Block validation failed');
     expect(error.data).toStrictEqual({
-      'pages.0.blocks.0.actions.onTap': 'Unknown action type',
+      'pages.0.blocks.0.actions.onTap': 'Custom action “onTap” is unused',
     });
   });